feat: check if \'Zap\' is already installed
Thanks to Senpai-10 for PR #75

+++ install.zsh
@@ -5,6 +5,10 @@ main() {
     local ZSHRC="${ZDOTDIR:-$HOME}/.zshrc"
     [[ $1 == "--branch" || $1 == "-b" && -n $2 ]] && local BRANCH="$2"
 
+    # check if ZAP_DIR already exists
+    [[ -d "$ZAP_DIR" ]] && { echo "Zap is already installed in '$ZAP_DIR'!" && read -q "res?Reinstall Zap? [y/N] "; echo "" }
+    [[ "$res" == "y" ]] && { echo "Reinstalling Zap..." && rm -rf "$ZAP_DIR" } || return 0
+
     git clone -b "${BRANCH:-master}" https://github.com/zap-zsh/zap.git "$ZAP_DIR" > /dev/null 2>&1 || { echo "❌ Failed to install Zap" && return 2 }
     mkdir -p "$ZAP_DIR/plugins"
 
@@ -14,6 +18,7 @@ main() {
 [ -f "$HOME/.local/share/zap/zap.zsh" ] && source "$HOME/.local/share/zap/zap.zsh"' "$ZSHRC"
 	fi
 # @formatter:on
+
     echo " Zapped"
     return 0
 }
